+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.8.2] - 2024-11-13
+
+### Added
+
+- Added `portalProps` prop to `Select`, `MultiSelect`, `Toast`, `AlertDialog`, `SimpleAlertDialog`, `ModalDialog`, and `SimpleModalDialog` components
+
+### Updated
+
+- Updated `Dropdown`, `Popover`, and `Tooltip` to fix issue with `portalProps` overriding the `className` prop
+
 ## [2.8.1] - 2024-10-21
 
 ### Added